+++ packages/plugin-openapi/src/index.ts
@@ -1,0 +1,5 @@
+import Plugin from "./Plugin";
+export * from "./types";
+export * from "./openapi";
+
+export default Plugin;