+++ packages/plugin-openapi/test/fixtures/api-users-collection-from/collections.js
@@ -1,0 +1,6 @@
+module.exports = [
+  {
+    id: "base",
+    routes: ["get-books:success"],
+  },
+];