+++ packages/plugin-openapi/test/fixtures/api-users-collection/openapi/api.js
@@ -1,0 +1,11 @@
+const openApiDocument = require("../../../openapi/users");
+
+module.exports = [
+  {
+    basePath: "/api",
+    collection: {
+      id: "users",
+    },
+    document: openApiDocument,
+  },
+];