+++ packages/plugin-openapi/test/fixtures/empty-method/openapi/api.js
@@ -1,0 +1,16 @@
+const deepMerge = require("deepmerge");
+
+const openApiDocument = require("../../../openapi/users");
+
+module.exports = [
+  {
+    document: deepMerge(openApiDocument, {
+      paths: {
+        "/users": {
+          get: undefined,
+          post: undefined,
+        },
+      },
+    }),
+  },
+];