+++ packages/plugin-openapi/test/fixtures/refs-remote-server/routes/refs.js
@@ -1,0 +1,17 @@
+const path = require("path");
+
+module.exports = [
+  {
+    id: "static",
+    url: "/",
+    variants: [
+      {
+        id: "success",
+        type: "static",
+        options: {
+          path: path.resolve(__dirname, "..", "static"),
+        },
+      },
+    ],
+  },
+];